- Make it compile cleanly with MS VC++ too, which is completely anal in some   cases.

+++ ext/date/lib/datetime.c
@@ -16,7 +16,7 @@
    +----------------------------------------------------------------------+
  */
 
-/* $Id: datetime.c,v 1.3 2005/06/15 07:15:38 derick Exp $ */
+/* $Id: datetime.c,v 1.4 2005/06/15 11:01:04 derick Exp $ */
 
 #include "timelib_structs.h"
 #include "datetime.h"
@@ -38,7 +38,7 @@ timelib_time* timelib_time_ctor()
 
 void timelib_time_tz_abbr_update(timelib_time* tm, char* tz_abbr)
 {
-	int i;
+	unsigned int i;
 	
 	TIMELIB_TIME_FREE(tm->tz_abbr);
 	tm->tz_abbr = strdup(tz_abbr);
@@ -141,7 +141,7 @@ signed long timelib_date_to_int(timelib_time *d, int *error)
 	if (error) {
 		*error = 0;
 	}
-	return d->sse;
+	return (signed long) d->sse;
 }
 
 void timelib_dump_date(timelib_time *d, int options)
